+++ src/main/java/com/wzw/work/service/KityMindService.java
@@ -6,7 +6,7 @@ import com.wzw.work.entity.KityMind;
  * @author Created by wuzhangwei on 2018/7/2921:16
  * @Description: TODO
  */
-public interface KityMindService extends BaseService<KityMind,Long>{
+public interface KityMindService extends BaseService<KityMind,Integer>{
 
 
 }
